requirements.txt
dnspython
async_dns
curio
gevent
